move email logic to email_notification_form contrib app
