+++ aldryn_forms/contrib/email_notification_form/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
