+++ aldryn_forms/contrib/email_notification_form/cms_plugin.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
